// types/akumina-core/interfaces/IPageWidget.ts
export default interface IPageWidget {
    Title: string;
    PageId: string;
    LayoutId: string;
    ContainerId: string;
    ZoneId: string;
    WidgetInstanceId: string;
    DisplayOrder: string;
    Grid: string;
    WidgetProperties: string;
    WidgetName: string;
    WidgetIcon: string;
    WidgetDescription: string;
    VirtualWidgetInstanceId: string;
    WidgetOptions: string;
}